backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2_cffi==25.1.0
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
openai==1.3.7
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.1.0
azure-core==1.35.1
azure-common==1.1.28
azure-mgmt-core==1.6.0
psycopg2-binary==2.9.9
email-validator==2.1.0
Jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
prometheus-client==0.19.0
psutil==5.9.6
PyGithub==1.59.1
github3.py==4.0.1
# Fix for missing modules
typing==3.7.4.3
# Additional dependencies
typing-extensions==4.8.0
anyio>=3.7.1,<4.0.0
starlette==0.27.0
click==8.1.7
h11==0.14.0
idna==3.4
sniffio==1.3.0
certifi==2023.7.22
charset-normalizer==3.3.2
urllib3==2.0.7
requests==2.31.0
aiofiles==23.2.1
watchfiles==0.21.0
python-dateutil==2.8.2
six==1.16.0
markupsafe==2.1.3
cryptography==41.0.7
ecdsa==0.18.0
pyasn1==0.5.0
rsa==4.9
bcrypt==4.1.2
argon2-cffi-bindings==21.2.0
cffi==1.16.0
pycparser==2.21
kombu==5.3.4
billiard==4.2.0
vine==5.1.0
tqdm==4.66.1
distro==1.8.0
jmespath==1.0.1
botocore==1.34.0
s3transfer==0.10.0
msal==1.25.0
msal-extensions==1.0.0
portalocker==2.8.2
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
requests-oauthlib==1.3.1
PyJWT==2.8.0
deprecated==1.2.14
wrapt==1.16.0
greenlet==3.0.1